Refactor Call classes
